Moved image reader stage from examples to stages, fixes to async loop and refactor process_sync method
