+++ ray-curator/ray_curator/stages/image/io/multi_source_image_reader.py
